fix(train): cast q/k back to v dtype after norm for FlashAttention
torch 2.13 added rms_norm to the autocast fp32 cast policy (pytorch/pytorch#174824),
so norm(q)/norm(k) now return fp32 under bf16 autocast instead of following the
input dtype like before. FlashAttention requires q/k/v to share one of
fp16/bf16/fp8 dtype, so the fp32 q/k broke the kernel call after the torch
2.13 upgrade in 5d8dbfc.

+++ train.py
@@ -88,7 +88,7 @@ class CausalSelfAttention(nn.Module):
 
         cos, sin = cos_sin
         q, k = apply_rotary_emb(q, cos, sin), apply_rotary_emb(k, cos, sin)
-        q, k = norm(q), norm(k)
+        q, k = norm(q).to(v.dtype), norm(k).to(v.dtype)
 
         y = fa3.flash_attn_func(q, k, v, causal=True, window_size=window_size)
         y = y.contiguous().view(B, T, -1)
